Apply persistence review fixes; add spec
Adversarial review (2-lens vs SQLDelight source): the ~200-line hand-typed DDL
verified CLEAN — all 9 tables, 7 triggers, 3 views, indexes, defaults, and FKs
byte-identical to source. Two minor adapter/mapper items fixed:

- UpdateStrategy decode now narrows to 32 bits (Int32(truncatingIfNeeded:)) to
  match Kotlin's toInt() for out-of-band stored values. Regression-tested.
- View/relation mappers (mapLibraryManga, mapMangaWithChapterCount,
  mapHistoryWithRelations) explicitly noted as deferred-with-views in Mappers.swift.

Spec: docs/specs/2026-07-21-persistence-layer-design.md. 106 tests green.

+++ Sources/MihonData/ColumnAdapters.swift
@@ -29,7 +29,9 @@ public enum StringListColumnAdapter {
 /// `.alwaysUpdate`, mirroring Kotlin's `getOrElse { ALWAYS_UPDATE }`.
 public enum UpdateStrategyColumnAdapter {
     public static func decode(_ dbValue: Int64) -> UpdateStrategy {
-        UpdateStrategy(rawValue: Int(truncatingIfNeeded: dbValue)) ?? .alwaysUpdate
+        // Narrow to 32 bits first to match Kotlin's `databaseValue.toInt()`
+        // before the enum lookup (they differ only for out-of-band stored values).
+        UpdateStrategy(rawValue: Int(Int32(truncatingIfNeeded: dbValue))) ?? .alwaysUpdate
     }
 
     public static func encode(_ value: UpdateStrategy) -> Int64 {
